+++ src/model_editor_update_texturing.cpp
@@ -209,7 +209,7 @@ void ModelEditor::_updateTexturingMenu()
 
 		if((_gui->getOverlay()->getValueFormId() == "textureRepeat") && _gui->getOverlay()->isValueFormConfirmed())
 		{
-			const auto content = static_cast<unsigned int>(max(0, Tools::parseInteger(_gui->getOverlay()->getValueFormContent())));
+			const auto content = max(0, Tools::parseInteger(_gui->getOverlay()->getValueFormContent()));
 
 			_fe3d->model_setTextureRepeat(_currentModelId, _currentPartId, textureRepeat);
 		}
