+++ src/quad2d_editor_update_menu.cpp
@@ -160,7 +160,7 @@ void Quad2dEditor::_updateChoiceMenu()
 		}
 		if((_gui->getOverlay()->getValueFormId() == "textureRepeat") && _gui->getOverlay()->isValueFormConfirmed())
 		{
-			const auto content = static_cast<unsigned int>(max(0, Tools::parseInteger(_gui->getOverlay()->getValueFormContent())));
+			const auto content = max(0, Tools::parseInteger(_gui->getOverlay()->getValueFormContent()));
 
 			_fe3d->quad2d_setTextureRepeat(_currentQuadId, content);
 		}
