+++ src/script_interpreter_func_fe3d_quad3d_set.cpp
@@ -348,7 +348,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dSetter(const string & functionNa
 		{
 			if(_validateFe3dQuad3d(args[0]->getString(), false))
 			{
-				_fe3d->quad3d_setTextureRepeat(args[0]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
+				_fe3d->quad3d_setTextureRepeat(args[0]->getString(), args[1]->getInteger());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
@@ -694,7 +694,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dSetter(const string & functionNa
 					return true;
 				}
 
-				_fe3d->quad3d_setAnimationRowIndex(args[0]->getString(), args[1]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
+				_fe3d->quad3d_setAnimationRowIndex(args[0]->getString(), args[1]->getString(), args[1]->getInteger());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
@@ -714,7 +714,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dSetter(const string & functionNa
 					return true;
 				}
 
-				_fe3d->quad3d_setAnimationColumnIndex(args[0]->getString(), args[1]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
+				_fe3d->quad3d_setAnimationColumnIndex(args[0]->getString(), args[1]->getString(), args[1]->getInteger());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
@@ -734,7 +734,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dSetter(const string & functionNa
 					return true;
 				}
 
-				_fe3d->quad3d_setAnimationIntervalMultiplier(args[0]->getString(), args[1]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
+				_fe3d->quad3d_setAnimationIntervalMultiplier(args[0]->getString(), args[1]->getString(), args[1]->getInteger());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
@@ -754,7 +754,7 @@ const bool ScriptInterpreter::_executeFe3dQuad3dSetter(const string & functionNa
 					return true;
 				}
 
-				_fe3d->quad3d_setAnimationIntervalDivider(args[0]->getString(), args[1]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
+				_fe3d->quad3d_setAnimationIntervalDivider(args[0]->getString(), args[1]->getString(), args[1]->getInteger());
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
