+++ src/water_editor_update_texturing.cpp
@@ -131,7 +131,7 @@ void WaterEditor::_updateTexturingMenu()
 
 		if((_gui->getOverlay()->getValueFormId() == "textureRepeat") && _gui->getOverlay()->isValueFormConfirmed())
 		{
-			const auto content = static_cast<unsigned int>(max(0, Tools::parseInteger(_gui->getOverlay()->getValueFormContent())));
+			const auto content = max(0, Tools::parseInteger(_gui->getOverlay()->getValueFormContent()));
 
 			_fe3d->water_setTextureRepeat(_currentWaterId, content);
 		}
